+++ backend/tests/Designer.Tests/Controllers/PolicyControllerTests/UpdateResourcePolicyTests.cs
@@ -1,0 +1,43 @@
+﻿using System.Net.Http;
+using System.Text;
+using System.Threading.Tasks;
+using Altinn.Studio.Designer.Controllers;
+using Altinn.Studio.PolicyAdmin.Models;
+using Designer.Tests.Controllers.ApiTests;
+using Designer.Tests.Utils;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Newtonsoft.Json;
+using Xunit;
+
+namespace Designer.Tests.Controllers.PolicyControllerTests
+{
+    public class UpdateResourcePolicyTests : DisagnerEndpointsTestsBase<PolicyController, UpdateResourcePolicyTests>
+    {
+        private readonly string _versionPrefix = "designer/api";
+
+        public UpdateResourcePolicyTests(WebApplicationFactory<PolicyController> factory) : base(factory)
+        {
+        }
+
+        [Fact]
+        public async Task Update_ResourcePolicyOk()
+        {
+            var targetRepository = TestDataHelper.GenerateTestRepoName();
+            await CopyRepositoryForTest("ttd", "ttd-resources", "testUser", targetRepository);
+
+            ResourcePolicy resourcePolicy = TestPolicyHelper.GenerateTestPolicy("ttd", targetRepository, null);
+
+            string responseBody;
+            string dataPathWithData = $"{_versionPrefix}/ttd/{targetRepository}/policy/ttdres1";
+            using (HttpRequestMessage httpRequestMessage = new HttpRequestMessage(HttpMethod.Put, dataPathWithData))
+            {
+                httpRequestMessage.Content = new StringContent(JsonConvert.SerializeObject(resourcePolicy), Encoding.UTF8, "application/json");
+                HttpResponseMessage response = await HttpClient.Value.SendAsync(httpRequestMessage);
+                response.EnsureSuccessStatusCode();
+                responseBody = await response.Content.ReadAsStringAsync();
+            }
+
+            Assert.NotEmpty(responseBody);
+        }
+    }
+}
